+++ impacthack/src/PastMedications.js
@@ -5,6 +5,11 @@ import { NavLink } from 'react-router-dom';
 async function genMess() {
     sessionStorage.setItem("message", "Successfully signed out");
 }
+
+
+let filtered = [];
+let userIdString = '';
+
 class PastMedications extends Component {
     state = {
         data: [],
@@ -15,7 +20,7 @@ class PastMedications extends Component {
         prescribedDay: 0,
         prescribedYear: 0,
         instructions: '',
-        userID: null
+        userID: ''
 
     };
 
@@ -60,29 +65,58 @@ class PastMedications extends Component {
             prescribedMonth: prescribedMonth,
             prescribedDay: prescribedDay,
             prescribedYear: prescribedYear,
-            instructions: instructions
+            instructions: instructions,
+            userID: userIdString
 
         });
     };
 
 
-    deleteFromDB = (idTodelete) => {
-        let objIdToDelete = null;
-        this.state.data.forEach((dat) => {
-            if (dat._id == idTodelete) {
-                objIdToDelete = dat._id;
+    deleteFromDB = (idTodelete, deleteOrMove) => {
+        let message = '';
+        if(deleteOrMove == 'delete'){
+            message = "Are you sure you want to delete this medication?";
+        } else if (deleteOrMove == 'move'){
+            message = "Are you sure you want to move this medication to past medication?";
+        }
+        if(window.confirm(message)){
+            let objIdToDelete = null;
+            this.state.data.forEach((dat) => {
+                if (dat._id == idTodelete) {
+                    objIdToDelete = dat._id;
+                }
+            });
+
+            axios.delete('http://localhost:3001/api/deletePastMedicationData', {
+                data: {
+                    id: objIdToDelete,
+                },
+            });
+        }
+        window.location.reload();
+    };
+
+    filterData = () => {
+        for(let i = 0; i < this.state.data.length; i++){
+            if(this.state.data[i].userID === userIdString && this.isIncluded(this.state.data[i])){
+                filtered.push(this.state.data[i]);
             }
-        });
+        }
+    }
 
-        axios.delete('http://localhost:3001/api/deletePastMedicationData', {
-            data: {
-                id: objIdToDelete,
-            },
-        });
-    };
+    isIncluded = (object) => {
+        for(let i = 0; i < filtered.length; i++){
+            if(filtered[i]._id === object._id){
+                return false;
+            }
+        }
+        return true;
+    }
 
     render() {
         const { data } = this.state;
+        userIdString = window.location.href.substring(window.location.href.indexOf("#") + 1 + 1, window.location.href.indexOf("/", window.location.href.indexOf("#") + 1 + 1));
+        this.filterData();
         return (
 
             <div>
@@ -96,22 +130,19 @@ class PastMedications extends Component {
                     <li><NavLink to="/" onClick={genMess}>Sign Out</NavLink></li>
                 </ul>
                 <ul>
-                    {data.length <= 0 ? 'NO DB ENTRIES YET' : data.map((dat) => (
-                        <li style={{ padding: '10px' }} key={data._id}>
+                    {filtered.length <= 0 ? 'NO DB ENTRIES YET' : filtered.map((dat) => (
+                        <li style={{ padding: '10px' }} key={dat._id}>
                             <span style={{ color: 'gray' }}> id: </span> {dat._id} <br />
                             <span style={{ color: 'gray' }}> Medication: </span> {dat.name} <br />
                             <span style={{ color: 'gray' }}> Type: </span> {dat.type} <br />
                             <span style={{ color: 'gray' }}> Prescribed Date: </span> {dat.prescribedMonth}/{dat.prescribedDay}/{dat.prescribedYear} <br />
                             <span style={{ color: 'gray' }}> Instructions: </span> {dat.instructions} <br />
-                            <button onClick={() => console.log("Edit Not Implemented Yet")}>
-                                EDIT - Not Implemented Yet :(
-                                </button> <br />
-                            <button onClick={() => this.deleteFromDB(dat._id)}>
+                            <button onClick={() => this.deleteFromDB(dat._id, 'delete')}>
                                 DELETE
                                 </button> <br />
                             <button onClick={() => {
                                 this.putCurrentDataToDB(dat.name, dat.type, dat.prescribedMonth, dat.prescribedDay, dat.prescribedYear, dat.instructions);
-                                this.deleteFromDB(dat._id);
+                                this.deleteFromDB(dat._id, 'move');
                             }}>
                                 RETURN TO CURRENT MEDICATION
                                 </button> <br />
